Add files via upload
In this recruitment competition, job seekers are provided with historical sales details of 45
Walmart stores located in various regions. Each store has multiple departments, and the task is to
predict the sales of each department in each store. In addition to the challenge, selected holiday
booking events are included in the database. These holidays affect sales, but it's a challenge to find
out which departments are affected and what's the magnitude of the impact.